+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Change Log
 
 ### master (untagged)
 
+* Add dynamic color
+
 ### 0.0.0 (2017-12-19)
 
 * Initial release